Splitting lib into separate files
To make it easier to work with the code and get an overview of each
part of the system. Also opens up more ideas for refactoring.
